add cppdoc (and fix a warning in ogl_shader)
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@6204 3db68df2-c116-0410-a063-a993310a9797

+++ source/lib/res/graphics/ogl_shader.cpp
@@ -115,7 +115,7 @@ static LibError Ogl_Shader_reload(Ogl_Shader* shdr, const VfsPath& pathname, Han
 	
 	{
 		const GLchar* strings[] = { (const GLchar*)file.get() };
-		const GLint tmp = file_size;
+		const GLint tmp = (GLint)file_size;
 		pglShaderSourceARB(shdr->id, 1, strings, &tmp);
 		pglCompileShaderARB(shdr->id);
 	}
